refactor: Put CommandRunner in package telemetry so telemetry compiles
Increase jest timeout to 15000 in findCodeObjects.spec.ts since it
fails consistently locally when all cli tests run.

In commandRunner.ts:run, change `err['code']` reference to `(err as
NodeJS.ErrnoException).code` because it didn't compile. Strange this
typecheck passed when commandRunner was in `cli`. The error was:

error TS7053: Element implicitly has an 'any' type because expression of type '"code"' can't be used to index type 'Error'.
