+++ packages/scanner/src/commandRunner.ts
@@ -1,0 +1,1 @@
+../../telemetry/src/commandRunner.ts
