+++ packages/scanner/src/commandStruct.ts
@@ -1,0 +1,1 @@
+../../telemetry/src/commandStruct.ts
