+++ packages/scanner/src/telemetry.ts
@@ -1,1 +1,1 @@
-../../telemetry/telemetry.ts
+../../telemetry/src/telemetry.ts
